+++ flutter_inappwebview_windows/windows/in_app_webview/in_app_webview.h
@@ -21,6 +21,8 @@
 #include "user_content_controller.h"
 #include "webview_channel_delegate.h"
 
+#include <WebView2EnvironmentOptions.h>
+
 namespace flutter_inappwebview_plugin
 {
   class InAppBrowser;