+++ flutter_inappwebview_windows/windows/in_app_webview/in_app_webview_settings.h
@@ -13,9 +13,9 @@ namespace flutter_inappwebview_plugin
     bool useShouldOverrideUrlLoading = false;
     bool useOnLoadResource = false;
     bool useOnDownloadStart = false;
+    bool useShouldInterceptRequest = false;
     std::string userAgent;
     bool javaScriptEnabled = true;
-    std::vector<std::string> resourceCustomSchemes;
     bool transparentBackground = false;
     bool supportZoom = true;
     bool isInspectable = true;