+++ flutter_inappwebview_windows/windows/in_app_webview/webview_channel_delegate.h
@@ -7,6 +7,7 @@
 #include "../types/base_callback_result.h"
 #include "../types/channel_delegate.h"
 #include "../types/create_window_action.h"
+#include "../types/custom_scheme_response.h"
 #include "../types/navigation_action.h"
 #include "../types/permission_response.h"
 #include "../types/web_resource_error.h"
@@ -48,6 +49,18 @@ namespace flutter_inappwebview_plugin
       ~PermissionRequestCallback() = default;
     };
 
+    class ShouldInterceptRequestCallback : public BaseCallbackResult<const std::shared_ptr<WebResourceResponse>> {
+    public:
+      ShouldInterceptRequestCallback();
+      ~ShouldInterceptRequestCallback() = default;
+    };
+
+    class LoadResourceWithCustomSchemeCallback : public BaseCallbackResult<const std::shared_ptr<CustomSchemeResponse>> {
+    public:
+      LoadResourceWithCustomSchemeCallback();
+      ~LoadResourceWithCustomSchemeCallback() = default;
+    };
+
     WebViewChannelDelegate(InAppWebView* webView, flutter::BinaryMessenger* messenger);
     WebViewChannelDelegate(InAppWebView* webView, flutter::BinaryMessenger* messenger, const std::string& name);
     ~WebViewChannelDelegate();
@@ -70,6 +83,8 @@ namespace flutter_inappwebview_plugin
     void onCreateWindow(std::shared_ptr<CreateWindowAction> createWindowAction, std::unique_ptr<CreateWindowCallback> callback) const;
     void onCloseWindow() const;
     void onPermissionRequest(const std::string& origin, const std::vector<int64_t>& resources, std::unique_ptr<PermissionRequestCallback> callback) const;
+    void shouldInterceptRequest(std::shared_ptr<WebResourceRequest> request, std::unique_ptr<ShouldInterceptRequestCallback> callback) const;
+    void onLoadResourceWithCustomScheme(std::shared_ptr<WebResourceRequest> request, std::unique_ptr<LoadResourceWithCustomSchemeCallback> callback) const;
   };
 }